// src/proto-ts/build/soong/linkerconfig/proto/linker_config.ts
export interface LinkerConfig {
  /** Extra permitted paths */
  permittedPaths: string[];
  /** Force APEX namespace visible */
  visible: boolean;
  /** Providing libs from the module */
  provideLibs: string[];
  /** Required libs from the module */
  requireLibs: string[];
  /** APEX can contribute search paths to specified namespaces. */
  contributions: LinkerConfig_Contribution[];
}

export interface LinkerConfig_Contribution {
  /** Target namespace where this module contributes the search paths. */
  namespace: string;
  /**
   * Search paths (inc. permitted paths) that this module contributes.
   * Paths should be related to the current module and can use "${LIB}" variable which is
   * expanded to "lib" or "lib64".
   * e.g. ${LIB}/subdir
   */
  paths: string[];
}

function createBaseLinkerConfig(): LinkerConfig {
  return { permittedPaths: [], visible: false, provideLibs: [], requireLibs: [], contributions: [] };
}

export const LinkerConfig: MessageFns<LinkerConfig> = {
  encode(message: LinkerConfig, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.permittedPaths) {
      writer.uint32(10).string(v!);
    }
    if (message.visible !== false) {
      writer.uint32(16).bool(message.visible);
    }
    for (const v of message.provideLibs) {
      writer.uint32(26).string(v!);
    }
    for (const v of message.requireLibs) {
      writer.uint32(34).string(v!);
    }
    for (const v of message.contributions) {
      LinkerConfig_Contribution.encode(v!, writer.uint32(42).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LinkerConfig {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLinkerConfig();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.permittedPaths.push(reader.string());
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.visible = reader.bool();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.provideLibs.push(reader.string());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.requireLibs.push(reader.string());
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.contributions.push(LinkerConfig_Contribution.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LinkerConfig {
    return {
      permittedPaths: globalThis.Array.isArray(object?.permittedPaths)
        ? object.permittedPaths.map((e: any) => globalThis.String(e))
        : [],
      visible: isSet(object.visible) ? globalThis.Boolean(object.visible) : false,
      provideLibs: globalThis.Array.isArray(object?.provideLibs)
        ? object.provideLibs.map((e: any) => globalThis.String(e))
        : [],
      requireLibs: globalThis.Array.isArray(object?.requireLibs)
        ? object.requireLibs.map((e: any) => globalThis.String(e))
        : [],
      contributions: globalThis.Array.isArray(object?.contributions)
        ? object.contributions.map((e: any) => LinkerConfig_Contribution.fromJSON(e))
        : [],
    };
  },

  toJSON(message: LinkerConfig): unknown {
    const obj: any = {};
    if (message.permittedPaths?.length) {
      obj.permittedPaths = message.permittedPaths;
    }
    if (message.visible !== false) {
      obj.visible = message.visible;
    }
    if (message.provideLibs?.length) {
      obj.provideLibs = message.provideLibs;
    }
    if (message.requireLibs?.length) {
      obj.requireLibs = message.requireLibs;
    }
    if (message.contributions?.length) {
      obj.contributions = message.contributions.map((e) => LinkerConfig_Contribution.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LinkerConfig>, I>>(base?: I): LinkerConfig {
    return LinkerConfig.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LinkerConfig>, I>>(object: I): LinkerConfig {
    const message = createBaseLinkerConfig();
    message.permittedPaths = object.permittedPaths?.map((e) => e) || [];
    message.visible = object.visible ?? false;
    message.provideLibs = object.provideLibs?.map((e) => e) || [];
    message.requireLibs = object.requireLibs?.map((e) => e) || [];
    message.contributions = object.contributions?.map((e) => LinkerConfig_Contribution.fromPartial(e)) || [];
    return message;
  },
};

function createBaseLinkerConfig_Contribution(): LinkerConfig_Contribution {
  return { namespace: "", paths: [] };
}

export const LinkerConfig_Contribution: MessageFns<LinkerConfig_Contribution> = {
  encode(message: LinkerConfig_Contribution, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.namespace !== "") {
      writer.uint32(10).string(message.namespace);
    }
    for (const v of message.paths) {
      writer.uint32(18).string(v!);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LinkerConfig_Contribution {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLinkerConfig_Contribution();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.namespace = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.paths.push(reader.string());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LinkerConfig_Contribution {
    return {
      namespace: isSet(object.namespace) ? globalThis.String(object.namespace) : "",
      paths: globalThis.Array.isArray(object?.paths) ? object.paths.map((e: any) => globalThis.String(e)) : [],
    };
  },

  toJSON(message: LinkerConfig_Contribution): unknown {
    const obj: any = {};
    if (message.namespace !== "") {
      obj.namespace = message.namespace;
    }
    if (message.paths?.length) {
      obj.paths = message.paths;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LinkerConfig_Contribution>, I>>(base?: I): LinkerConfig_Contribution {
    return LinkerConfig_Contribution.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LinkerConfig_Contribution>, I>>(object: I): LinkerConfig_Contribution {
    const message = createBaseLinkerConfig_Contribution();
    message.namespace = object.namespace ?? "";
    message.paths = object.paths?.map((e) => e) || [];
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}